LayoutTests/networkinformation/multiple-frames-expected.txt

Tests using NetworkInfo from multiple frames.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS connection.bandwidth is bandwidth
PASS connection.metered is metered
PASS connection.bandwidth is bandwidth
PASS connection.metered is metered
PASS successfullyParsed is true

TEST COMPLETE
